Add "repeat mode" for the Media_Next key
As I tend to skip tracks in by pressing Media_Next repeatedly, this commit changes the way that key is handled.

Watching for a change to the reg key happens inside the handler instead of going all the way back to the start. If a different key is pressed or 500ms is reached, the loop is broken and control goes back to the main loop handling all the keys. If Media_Next is pressed again, the handler gets control again.

The downside to this is that the desktop change event being signalled or WTS's notification for a session change may get discarded, knocking the inScriptSession and onScriptDesktop variables out of sync. If it is actually a problem, then those variables' states can get manually updated after the WaitForSingleObject or MsgWaitForMultipleObjectsEx can be used again.